# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kouvee Petshop API Documentation",
        "contact": {
            "name": "Natan Hari P.",
            "email": "natanhp@protonmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/customers/restore/{id}": {
            "get": {
                "tags": [
                    "customers"
                ],
                "description": "Restore the delted customer",
                "operationId": "App\\Http\\Controllers\\CustomersController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of customer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted customer"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/employees/restore/{id}": {
            "get": {
                "tags": [
                    "employees"
                ],
                "description": "Restore the delted employee",
                "operationId": "App\\Http\\Controllers\\EmployeesController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of employee",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted employee"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/petsizes/restore/{id}": {
            "get": {
                "tags": [
                    "pet sizes"
                ],
                "description": "Restore the deleted pet size",
                "operationId": "App\\Http\\Controllers\\PetSizesController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a pet size",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted pet size"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/uni/pettypes/getall": {
            "get": {
                "tags": [
                    "pet types"
                ],
                "description": "Get all the pet types",
                "operationId": "App\\Http\\Controllers\\PetTypesController::getAll",
                "responses": {
                    "default": {
                        "description": "Get all the pet types"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/pettypes/insert": {
            "post": {
                "tags": [
                    "pet types"
                ],
                "summary": "Insert a new service",
                "operationId": "App\\Http\\Controllers\\PetTypesController::insert",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "description": "The type of the pet",
                                        "type": "string"
                                    },
                                    "createdBy": {
                                        "description": "The id of the owner who creates the pet type",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/uni/pettypes/getbyid/{id}": {
            "get": {
                "tags": [
                    "pet types"
                ],
                "description": "Get a pet type by id",
                "operationId": "App\\Http\\Controllers\\PetTypesController::getPetTypeById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the pet type",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Get a pet type by id"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/uni/pettypes/getbytype/{type}": {
            "get": {
                "tags": [
                    "pet types"
                ],
                "description": "Get a pet type by type",
                "operationId": "App\\Http\\Controllers\\PetTypesController::getPetTypeByType",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Type of the pet",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Get a pet type by type"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/pettypes/update": {
            "put": {
                "tags": [
                    "pet types"
                ],
                "summary": "Update a pet type",
                "operationId": "App\\Http\\Controllers\\PetTypesController::update",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "The id of the pet type",
                                        "type": "integer"
                                    },
                                    "type": {
                                        "description": "The type of the pet",
                                        "type": "string"
                                    },
                                    "updatedBy": {
                                        "description": "The foreign key of the owner who updates the pet type",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/pettypes/delete/{id}/{ownerId}": {
            "delete": {
                "tags": [
                    "pet types"
                ],
                "summary": "Deletes a pet type",
                "operationId": "App\\Http\\Controllers\\PetTypesController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Pet type id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "ownerId",
                        "in": "path",
                        "description": "Owner who deletes the pet type",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Pet type not deleted it's because either the deletion failed or Pet type to be deleted not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/pettypes/restore/{id}": {
            "get": {
                "tags": [
                    "pet types"
                ],
                "description": "Restore the deleted pet type",
                "operationId": "App\\Http\\Controllers\\PetTypesController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a pet type",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted pet type"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/pets/restore/{id}": {
            "get": {
                "tags": [
                    "pets"
                ],
                "description": "Restore the deleted pet",
                "operationId": "App\\Http\\Controllers\\PetsController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a pet",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted pet"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/products/restore/{id}": {
            "get": {
                "tags": [
                    "products"
                ],
                "description": "Restore the delted product",
                "operationId": "App\\Http\\Controllers\\ProductsController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of product",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted product"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/servicedetails/restore/{id}": {
            "get": {
                "tags": [
                    "service details"
                ],
                "description": "Restore the deleted service detail",
                "operationId": "App\\Http\\Controllers\\ServiceDetailsController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a service detail",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted service detail"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/services/restore/{id}": {
            "get": {
                "tags": [
                    "services"
                ],
                "description": "Restore the deleted service",
                "operationId": "App\\Http\\Controllers\\ServicesController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a service",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted service"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/suppliers/restore/{id}": {
            "get": {
                "tags": [
                    "suppliers"
                ],
                "description": "Restore the delted supplier",
                "operationId": "App\\Http\\Controllers\\SuppliersController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of supplier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restored the deleted supplier"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}